% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcsurg}
\alias{bcsurg}
\title{Data to showcase differences in overall survival among subgroups of breast cancer survivors}
\format{
A data frame with 1598 rows and 31 variables
 \describe{
.     \item{ID..}{Subject ID}
     \item{Groups}{5 groups of breast cancer}
     \item{Breast_surgery_code}{Surgery type}
     \item{DT_dxdate2}{Date of cancer diagnosis}
     \item{DT_dod}{Date of death}
     \item{DT_date_last_seen}{Date of last follow-up}
     }
}
\usage{
bcsurg
}
\description{
Contains the grouping variable, and essential information needed to derive the overall survival time and status of each breast cancer survivor.
}
\examples{
data(bcsurg)
}
\keyword{datasets}
