% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survcomp.R
\name{survcomp}
\alias{survcomp}
\title{Post-hoc Within Group Comparisons of Differences in a Survival Outcome}
\usage{
survcomp(dat, patid, x_c, x_n, dt_start, dt_outcome, dt_end, unit = "month")
}
\arguments{
\item{dat}{a data frame.}

\item{patid}{an integer vector of the subject IDs.}

\item{x_c}{the name(s) of categorical variable(s).}

\item{x_n}{the name(s) of numeric variable(s).}

\item{dt_start}{the name of the date of the starting point.}

\item{dt_outcome}{the name of the date of the survival outcome.}

\item{dt_end}{the name of the date of the endpoint.}

\item{unit}{the unit of measurement for the time variable; options are: "day", "month", "year".}
}
\value{
a summary_emm object and a data frame of pairwise comparisons, confidence intervals, test statistics, and p-values.
}
\description{
Performs pairwise comparisons of the levels of one categorical variable within each level of another categorical variable adjusting for multiplicity.
}
\examples{
# Read in the sample data
data(dat_da)
emm_contrasts <- survcomp(
dat = dat_da, patid = "ID..",
x_c = c("Groups", "Breast_surgery_code"),
x_n = "AGE_agedx", dt_start = "DT_dxdate2",
dt_outcome = "DT_dod",
dt_end = "DT_date_last_seen"
)
# Only keep the statistically significant pairwise comparisons
emm_contrasts[emm_contrasts$p.value < 0.05, ]
}
