% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survcomp_plot.R
\name{survcomp_plot}
\alias{survcomp_plot}
\title{Kapan-Meier Plots and Log-rank Test}
\usage{
survcomp_plot(dat, patid, f1, f2, dt_start, dt_outcome, dt_end, unit = "month")
}
\arguments{
\item{dat}{a data frame.}

\item{patid}{an integer vector of the subject IDs.}

\item{f1}{the name of the first categorical variable.}

\item{f2}{the name of the second categorical variable.}

\item{dt_start}{the name of the date of the starting point.}

\item{dt_outcome}{the name of the date of the survival outcome.}

\item{dt_end}{the name of the date of the endpoint.}

\item{unit}{the unit of measurement for the time variable; options are: "day", "month", "year".}
}
\value{
Multiple Kaplan-Meier plots with a display of the p-values from log-rank tests.
}
\description{
Draws Kaplan-Meier plots of the survival outcome of interest, and compares the levels of one categorical variable within each level of another categorical variable globally via log-rank tests.
}
